Updated architecture according to #56
