+++ R3E.Relay/Program.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
-using R3E.API;
+using R3E.Core.SharedMemory;
 using R3E.Data;
 using R3E.UdpRelay;
 using System.Net;
