+++ R3E.YaHud/Components/Pages/Index.razor
@@ -5,13 +5,9 @@
 @inject VisibilityService VisibilityService
 @inject TestModeService TestModeService
 @inject SettingsService SettingsService
-@using R3E.API
-@using R3E.YaHud.Components.Widget.LapProgress
-@using R3E.YaHud.Components.Widget.RelativeDriver
 @using R3E.YaHud.Services
 @using R3E.YaHud.Services.Settings
 @using R3E.YaHud.Components.UI.Components
-@using R3E.YaHud.Components.Widget
 
 @if (!locked)
 {
