+++ R3E.YaHud/Components/Widget/FuelCalc/FuelCalc.razor
@@ -1,6 +1,5 @@
-﻿@using R3E.Data
+﻿@using R3E.Features.Fuel
 @using R3E.YaHud.Components.Widget.Core
-@using R3E.Models
 @inherits HudWidgetBase<FuelSettings>
 
 @if (Settings?.Visible ?? false)
