+++ R3E.YaHud/Components/Widget/LapProgress/LapProgress.razor
@@ -1,7 +1,7 @@
-﻿@using R3E.API
+﻿@using System.Globalization
+@using R3E.Core.Services
 @using R3E.YaHud.Components.Widget.Core
 @using R3E.YaHud.Components.Widget.LapProgress.Components
-@using System.Globalization
 @inherits HudWidgetBase<LapProgressSettings>
 @implements IDisposable
