+++ R3E.YaHud/Components/Widget/MoTec/MoTec.razor
@@ -1,6 +1,6 @@
-﻿@using R3E.YaHud.Components.Widget.Core
+﻿@using R3E.Converters
+@using R3E.YaHud.Components.Widget.Core
 @using R3E.YaHud.Components.Widget.MoTec.Components
-@using R3E.API
 @inherits HudWidgetBase<MoTecSettings>
 
 @if (Settings?.Visible ?? false)
