+++ R3E.YaHud/Components/Widget/RelativeDriver/RelativeDriver.razor
@@ -1,9 +1,11 @@
-﻿@using R3E.API
-@using R3E.API.Image
-@using R3E.API.TimeGap
+﻿@using R3E.Features.Image
+@using R3E.Features.TimeGap
+@using R3E.Core.Services
 @using R3E.YaHud.Components.Widget.Core
 @using R3E.YaHud.Components.Widget.RelativeDriver
 @using R3E.YaHud.Components.Widget.RelativeDriver.Models
+idget.RelativeDriver.Models
+get.RelativeDriver.Models
 @inherits HudWidgetBase<RelativeDriverSettings>
 @inject IImageService ImageService
 @inject ITimeGapService TimeGapService
