+++ R3E.YaHud/Components/Widget/StartLight/StartLight.razor
@@ -1,5 +1,4 @@
 ﻿@using R3E.YaHud.Components.Widget.Core
-@using R3E.API
 @inherits HudWidgetBase<StartLightSettings>
 
 @if (Settings?.Visible ?? false)
