+++ R3E.YaHud/Components/Widget/UserInputs/UserInputs.razor
@@ -1,6 +1,5 @@
 ﻿@using R3E.YaHud.Components.Widget.Core
 @using R3E.YaHud.Components.Widget.UserInputs.Components
-@using R3E.API
 @inherits HudWidgetBase<UserInputsSettings>
 
 @if (Settings?.Visible ?? false)
