+++ R3E.YaHud/Services/Settings/GlobalSettings.cs
@@ -1,4 +1,4 @@
-﻿using R3E.API;
+﻿using R3E.Converters;
 
 namespace R3E.YaHud.Services.Settings
 {
