--- R3E/API/Image/ImageSize.cs
@@ -1,9 +1,0 @@
-namespace R3E.API.Image
-{
-    public enum ImageSize
-    {
-        Full,
-        Small,
-        Thumb
-    }
-}
