--- R3E/API/TimeGap/ITimeGapService.cs
@@ -1,7 +1,0 @@
-﻿namespace R3E.API.TimeGap
-{
-    public interface ITimeGapService
-    {
-        float GetTimeGapRelative(int subjectSlotId, int targetSlotId);
-    }
-}
