--- R3E/API/Unit.cs
@@ -1,21 +1,0 @@
-﻿namespace R3E.API
-{
-    public enum SpeedUnit
-    {
-        MetersPerSecond,
-        KilometersPerHour,
-        MilesPerHour,
-    }
-
-    public enum AngularUnit
-    {
-        RadiansPerSecond,
-        RevolutionsPerMinute,
-    }
-
-    public enum TemperatureUnit
-    {
-        Celsius,
-        Fahrenheit,
-    }
-}
