+++ R3E/Converters/AngularUnit.cs
@@ -1,0 +1,8 @@
+﻿namespace R3E.Converters
+{
+    public enum AngularUnit
+    {
+        RadiansPerSecond,
+        RevolutionsPerMinute,
+    }
+}
