+++ R3E/Converters/SpeedUnit.cs
@@ -1,0 +1,9 @@
+﻿namespace R3E.Converters
+{
+    public enum SpeedUnit
+    {
+        MetersPerSecond,
+        KilometersPerHour,
+        MilesPerHour,
+    }
+}
