+++ R3E/Converters/TemperatureUnit.cs
@@ -1,0 +1,8 @@
+﻿namespace R3E.Converters
+{
+    public enum TemperatureUnit
+    {
+        Celsius,
+        Fahrenheit,
+    }
+}
