+++ R3E/Features/Image/ImageSize.cs
@@ -1,0 +1,9 @@
+namespace R3E.Features.Image
+{
+    public enum ImageSize
+    {
+        Full,
+        Small,
+        Thumb
+    }
+}
