+++ R3E/Features/TimeGap/ITimeGapService.cs
@@ -1,0 +1,7 @@
+﻿namespace R3E.Features.TimeGap
+{
+    public interface ITimeGapService
+    {
+        float GetTimeGapRelative(int subjectSlotId, int targetSlotId);
+    }
+}
